Mount procfs at /proc in the rootfs
In order to be able to use the rootfs as a proper container, we need
to disable setgroups syscall, which allows us to write into gid_map,
which allows us to assing ourselves to root group. Then we use setuid
and setgid syscalls to make sure that we are actually root. Then we
need to mount / as MS_PRIVATE, because by default it is mounted as
MS_SHARED (systemd does that). And then we can finally mount proc as
/proc, yay!

Signed-off-by: Aodhnait Étain <aodhneine@pm.me>